requirements-dev.txt
flake8==3.9.2;python_version<"3.6"
flake8==5.0.4;python_version>="3.6"
m2r2==0.3.2;python_version<"3.7"
m2r2==0.3.3.post2;python_version>="3.7"
numpydoc==0.9.2;python_version<"3.5"
numpydoc==1.1.0;python_version>="3.5" and python_version<"3.7"
numpydoc==1.5.0;python_version>="3.7"
plexapi-backport[alert]==4.15.2
pytest==4.6.11;python_version<"3"
pytest-cov==2.12.1;python_version<"3"
rstcheck==3.5.0;python_version<"3.7"
rstcheck==6.1.2;python_version>="3.7" and python_version<"3.8"
rstcheck==6.2.0;python_version>="3.8"
Sphinx==1.8.6;python_version<"3.5"
Sphinx==3.5.4;python_version>="3.5" and python_version<"3.6"
Sphinx==5.3.0;python_version>="3.6" and python_version<"3.8"
Sphinx==7.1.2;python_version>="3.8" and python_version<"3.9"
Sphinx==7.2.6;python_version>="3.9"
sphinx-rtd-theme==1.2.0
tqdm==4.64.1;python_version<"3"
